feat: add withdrawalMutation to useAuth hook
