+++ src/pages/accounts/edit/withdrawal/index.tsx
@@ -1,12 +1,10 @@
-import { postWithdrawUser } from "@/api/auth/withdrawal";
 import DeleteConfirmDialog from "@/components/@common/FeedCard/components/DeleteConfirmDialog";
 import FormButton from "@/components/@common/FormButton";
 import FormInput from "@/components/@common/FormInput";
 import FormLayout from "@/components/@common/FormLayout";
 import Logo from "@/components/@common/Logo";
-import { APIError } from "@/types/error";
-import { Center, useDisclosure, useToast } from "@chakra-ui/react";
-import { useMutation } from "@tanstack/react-query";
+import useAuth from "@/hooks/useAuth";
+import { Center, useDisclosure } from "@chakra-ui/react";
 import Head from "next/head";
 import { useRouter } from "next/router";
 import { useForm } from "react-hook-form";
@@ -17,7 +15,6 @@ type FormData = {
 
 export default function Withdrawal() {
   const router = useRouter();
-  const toast = useToast();
 
   const {
     isOpen: isAlertOpen,
@@ -27,27 +24,17 @@ export default function Withdrawal() {
 
   const {
     register,
-    getValues,
     formState: { errors, isSubmitting, isValid }
   } = useForm<FormData>({ mode: "onChange" });
 
-  const withdrawalMutation = useMutation({
-    mutationFn: ({ password }: { password: string }) =>
-      postWithdrawUser({ password }),
-    onSuccess: () => {
-      router.back();
-    },
-    onError: (e: APIError) => {
-      toast({
-        title:
-          e.errorCode === "WRONG_EMAIL_OR_PASSWORD_BAD_REQUEST"
-            ? "비밀번호가 올바르지 않아요"
-            : "회원 탈퇴에 실패했어요",
-        status: "error",
-        duration: 3000
-      });
-    }
-  });
+  const { withdrawalMutation } = useAuth();
+  const withdrawal = () => {
+    withdrawalMutation.mutate(undefined, {
+      onSuccess: () => {
+        router.push("/");
+      }
+    });
+  };
 
   return (
     <>
@@ -59,7 +46,7 @@ export default function Withdrawal() {
         isOpen={isAlertOpen}
         onClose={onAlertClose}
         onDelete={() => {
-          withdrawalMutation.mutate({ password: getValues("password") });
+          withdrawal();
           onAlertClose();
         }}
       />
